client/requirements.txt
cryptography==40.0.1
PyJWT==2.6.0
